requirements.txt
gradio
mysql-connector-python
openai
python-dotenv
pillow
